python_api/requirements.txt
fastapi
uvicorn
notebooklm-py
httpx[http2]
python-multipart
curl_cffi
playwrightpython-dotenv
python-pptx
pdfkit
python-docx
